Extra volumes support to manila-operator
This patch aligns the code with the lib-common ExtraVolumes change,
where the ExtraVolumes support is provided for each storage component.
Additional extra mounts can be defined in the top level manila CR and
passed to the GetVolume functions provided by lib-common. Then,
according to the Propagation policy, the extraMount will be (or not be)
propagated to the service deployed by manila-operator.

Depends-on: openstack-k8s-operators/lib-common#88

Signed-off-by: Francesco Pantano <fpantano@redhat.com>