Add service db finalizer logic
Podified operators which use service dbs and the db.CreateOrPatchDB
method , now will get the finalizer set on the db automatically when
using the newest lib-common, but they won't have the code remove it on
cleanup i.e reconcileDelete, which might make the delete operator action
hang. This adds the service Db removal code in the operator delete
(reconcileDelete) function to work with the above.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>