Rename master branch to main
We have issues deploying the manila-operator via openstack-operator
because the script which is supposed to run the github action that
pushes the new bundle has "master" instead of "main". This results in
having a "main-latest" [1][2][3] tag instead of just "latest" and the
meta-operator is not able to resolve the right manila container!

[1] https://quay.io/repository/openstack-k8s-operators/manila-operator?tab=tags
[2] https://quay.io/repository/openstack-k8s-operators/manila-operator-index?tab=tags
[3] https://quay.io/repository/openstack-k8s-operators/manila-operator-bundle?tab=tags

Signed-off-by: Francesco Pantano <fpantano@redhat.com>