Improve Healthckecks for manila-operator
Manila-operator is currently doing limited actions when it comes
to doing healthchecks. This change enhances the probes so we can
do actual checks on the status of the manila services, by
implementing a simple HTTP server (as being done by the
cinder operator)

The HTTP server though will do the healthchecks using a command we
offer in Manila: manila-manage service list. The HTTP server will
take as arguments the type of service you are trying to ensure is
alive (share or scheduler) and the location of the manila
configuration file.

Manila-manage service list command does not require any kind of
additional client to be run, nor keystone tokens or additional
authentication steps are required. The command will simply get the
services that are already registered in the database.

Then, the HTTP server will check on the state of the services and
respond to the operator according to the state and filters
matching.