Remove crudini dependency and unused scripts
This change is part of a series that aim to remove the crudini
dependency and be consistent with what we've done in Cinder PR#108. In
particular, secret snippets can be rendered in the config exposing
customServiceConfigSecrets and processed by the init script. The
customServiceConfigSecrets parameter is used to define a list of secrets
(by name) that are processed by the operator and the init script and
propagated to the manila services.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>