Enable webhooks in envTest and update the DefaultContainerImage test
This change improves the existing envTest starting the webhooks within
the suite_test setup and fixing the ContainerImage test accordingly.
In addition, as done for dataplane-operator [1], this patch improves the
Make target using ginkgo cli, while go test is still used to run regular
unit tests. This is very useful for development purposes, where we would
like to focus on that single test that is going to be written and get a
more verbose log, as well as being able to customize concurrency and
other fields.

'''
e.g.

make test GINKGO_ARGS="-v --focus \
  "'Should fail if db-sync job fails when DB is Created'"
'''

Other than that, it's easier to run tests in parallel and produce
reports [2].

[1] openstack-k8s-operators/dataplane-operator#204
[2] https://onsi.github.io/ginkgo/#the-ginkgo-cli-vs-go-test

Signed-off-by: Francesco Pantano <fpantano@redhat.com>