Improve NetworkAttachments check
Only verify networkAttachments if instance.Spec.Replicas > 0.
In addition, Network validation is checking the pods if IPs are correct
reported for each of the requested networks. For this a label selecor is
used to get only the pods for the sub manila component (api, scheduler
and share). Right now there is only the generic manila app label on the
deployment resources set. This adds a component label on all the api,
scheduler and share deployment resources which allows the check to only
get the wanted resource pods.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>