Do not log on /dev/stdout
Due to the recent failures with the new crc/sno versions, Pods are not
allowed to log directly in /dev/stdout. For this reason, like we've done
in glance-operator, this patch imports a minimal logging.conf that is
added in both manilaShare and manilaScheduler, while it's not required
by manilaAPI due to the nature of the service and the way it's started.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>