Recalculate deployment status
If we don't have any Manila instance in the deployment, we should mark
the deployment condition to False. If manila is enabled in the openstack
control plane, we always need at least 1 replica of API and scheduler,
while the manilaShare service can be rolled out later in time (according
to the backends).
As we've done in GlanceAPI [1], this patch adds the same behavior to the
Manila operator.

[1] https://github.com/openstack-k8s-operators/glance-operator/pull/311

Signed-off-by: Francesco Pantano <fpantano@redhat.com>