Add manila db purge job
Since Ocata release [1], Manila provides a utility command to purge deleted rows
from database [2], and this command can be executed periodically to avoid
database tables from getting bigger.
This patch introduces a Manila CronJob to periodically run manila-manage db
purge, so that soft-deleted records are removed from Manila database.
The CronScheduling is not exposed through the CR API but it's aligned
with the default values defined for TripleO based deployments.

[1] https://specs.openstack.org/openstack/manila-specs/specs/ocata/add-db-manage-purge.html
[2] https://docs.openstack.org/manila/latest/cli/manila-manage.html

Signed-off-by: Francesco Pantano <fpantano@redhat.com>