[go.mod]Bump from v0.3.0 to pseudoversion
Our intention is to track service operator and lib-common dependencies
via pseudoversions. However renovate does not automatically bump from a
tagged version (e.g. v0.3.0) to a newer but not tagged pseudoversion. So
this patch does the manual bump. After this renovate will bump the newer
pseduoversions.

Adapted to the new version by:

* adding RBAC rules for MariaDBAccount manipulation
* adding Own(MariaDBAccount) to detect status changes
* adapting test to simulate MariaDBAccount success
* adapting the tests to the interim changes of mariadb TLS