Build a FIPS compatible operator image
This change is aligned with the work already done for other operators,
and the goal is to build a FIPS compatible manila-operator image. For
this reason, the BASE_IMAGE has been moved to ubi-minimal and the
toolchain uses go-toolset. Both CGO and GO111MODULE are enabled and
passed as extra args to go build.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>