Update all conditions on every reconciliation
This PR follows recent changes to the cinder-operator, which
basically initializes conditions to Unknown at the start of the
reconcilation loop. As reconciliation progresses, conditions
are updated to True or False, and if reconciliation doesn't
complete then any remaining conditions are left Unknown.

The ReadyCondition is set True when reconciliation successfully
completes. For the api, scheduler and share services, this means
the number of replicas matches the requested value. Again, this
is consistent with the way cinder behaves.

For the top manila controller, ReadyCondition is set True when all
the other manila controllers are ready.