Fix api log file increases indefinitely
The manila-api pod uses a file between the manila-api service and a
sidecar to pass the manila logs from the service to the stdout.

With the current code it will grow forever, which makes no sense, since
the stdout from the pod itself is already being stored, under
/var/log/pods and/or in a centralized logging service.

This patch makes the oslo.log service rotate the log file once it
reaches 20MB in size.

Other projects like nova and glance set the limit to 50MB and they keep
5 rotated files, but I think that keeping rotated files is useless here
and 50MB seems a bit much.

Unfortunately we need to at least keep 1 copy, or oslo.log will not do
the right thing if set to 0.

We also need to redirect errors in the sidecar container that tails the
log to /dev/null to prevent tail error messages whenever the file gets
rotated.