Move to VerifySecret when checking the ctlplane secret and trasportURL
Currently we get the osp_secret and we build the hash of the entire
secret. If an unrelated field is added or changed, this cause a rollout
of new manila Pods, resultin in an entire service restarts in all its
components.
This change does two main things:
1. replace the GetSecret call to VerifySecret
2. build a common ensureSecret function that is called by all
   controllers (top level, manilaAPI, manilaScheduler, manilaShare)

By having a common file for all controllers we are able to also start
some code deduplication.

Jira: https://issues.redhat.com/browse/OSPRH-8290

Signed-off-by: Francesco Pantano <fpantano@redhat.com>