Add ManilaShare instance name as label
As done for Glance, where APIName might be more than one and have
arbitrary names, it's easier to store and retrieve the instance Name
from the label, instead of hardcoding part of the pod name in a string.
This way the code is more solid and we have to reduce the amount of
string manipulation.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>