Run manila service cleanup when a share is deleted
When a share is deleted from the top-level CR, its reference is not
deleted, and manila still sees it as Down/not available. For this
reason, every time a scale down operation is executed, we need to run a
job that is supposed to run a service cleanup, removing everything set
as down.

Jira: https://issues.redhat.com/browse/OSPRH-6526

Signed-off-by: Francesco Pantano <fpantano@redhat.com>