Run a cleanup Job when ManilaShare is scaled down
When a backend config is removed, the "service" database entry in manila needs
to be adjusted to account for the removal. This patch introduces a Job that
runs a manila-manage command every time a share is scaled down. The main job
definition is now more flexible and we can pass a TTL in case we want to
customize the time the job is kept.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>