Delay jon when a share is deleted
If the share cleanup job is executed right after the share deletion, it
might be possible that the other manila components have not yet been
notified about the event. The result is that the job has no effect.
This patch introduces a parameter for the ManilaJob that is supposed to
delay the command. A constant has been defined to properly indentify the
delay time associated with the service cleanup.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>