Validate common empty values on CR update
We recently hit an issue with the openstack control plane where
manila-operator gets stuck waiting for a transportURL without
progressing or reporting an issue.
During a CR update, it would be ideal to provide a feedback to
the human operator about the not admitted empty parameters, and
the goal of this patch is to improve the user experience by
introducing base params webhook validations.

Jira: https://issues.redhat.com/browse/OSPRH-16366

Signed-off-by: Francesco Pantano <fpantano@redhat.com>